Arm and lander action definitions moved into respective files
