--- ow_lander/src/ow_lander/actions/grind.py
@@ -1,37 +1,0 @@
-# The Notices and Disclaimers for Ocean Worlds Autonomy Testbed for Exploration
-# Research and Simulation can be found in README.md in the root directory of
-# this repository.
-
-import rospy
-
-import ow_lander.msg
-
-from ow_lander.server import ActionServerBase
-from ow_lander.arm_action_mixin import ArmActionMixin
-
-class GrindServer(ArmActionMixin, ActionServerBase):
-
-  name          = 'Grind'
-  action_type   = ow_lander.msg.GrindAction
-  goal_type     = ow_lander.msg.GrindGoal
-  feedback_type = ow_lander.msg.GrindFeedback
-  result_type   = ow_lander.msg.GrindResult
-
-  def __init__(self):
-    super().__init__()
-
-  def execute_action(self, goal):
-    try:
-      self._arm.checkout_arm(self.name)
-      self._arm.switch_to_grinder_controller()
-      plan = self._planner.grind(goal)
-      self._arm.execute_arm_trajectory(plan)
-    except RuntimeError as err:
-      self._set_aborted(str(err),
-        final=self._arm_tip_monitor.get_link_position())
-    else:
-      self._set_succeeded("Arm trajectory succeeded",
-        final=self._arm_tip_monitor.get_link_position())
-    finally:
-      self._arm.switch_to_arm_controller()
-      self._arm.checkin_arm(self.name)
