--- ow_lander/src/ow_lander/actions/stop.py
@@ -1,29 +1,0 @@
-# The Notices and Disclaimers for Ocean Worlds Autonomy Testbed for Exploration
-# Research and Simulation can be found in README.md in the root directory of
-# this repository.
-
-import rospy
-
-import ow_lander.msg
-
-from ow_lander.server import ActionServerBase
-from ow_lander.arm_action_mixin import ArmActionMixin
-
-class StopServer(ArmActionMixin, ActionServerBase):
-
-  name          = 'Stop'
-  action_type   = ow_lander.msg.StopAction
-  goal_type     = ow_lander.msg.StopGoal
-  feedback_type = ow_lander.msg.StopFeedback
-  result_type   = ow_lander.msg.StopResult
-
-  def __init__(self):
-    super().__init__()
-
-  def execute_action(self, _goal):
-    if self._arm.stop_arm():
-      self._set_succeeded("Arm trajectory stopped",
-        final=self._arm_tip_monitor.get_link_position())
-    else:
-      self._set_aborted("No arm trajectory to stop",
-        final=self._arm_tip_monitor.get_link_position())
